moves the MapContainer in a separate folder
