+++ src/components/MapContainer/reducer.js
